+++ apps/extension/src/core/handlers/stores.ts
@@ -3,8 +3,9 @@ import { AccountsCatalogData } from "@core/domains/accounts/store.catalog"
 import { AppStoreData, appStore } from "@core/domains/app/store.app"
 import { ErrorsStoreData, errorsStore } from "@core/domains/app/store.errors"
 import { PasswordStoreData, passwordStore } from "@core/domains/app/store.password"
-import { RemoteConfigStoreData, remoteConfigStore } from "@core/domains/app/store.remoteConfig"
+import { remoteConfigStore } from "@core/domains/app/store.remoteConfig"
 import { SettingsStoreData, settingsStore } from "@core/domains/app/store.settings"
+import { RemoteConfigStoreData } from "@core/domains/app/types"
 import { balanceStore } from "@core/domains/balances/store"
 import { MnemonicData, mnemonicsStore } from "@core/domains/mnemonics/store"
 import { sitesAuthorisationStore } from "@core/domains/sitesAuthorised"
